converted to double precision mathematics

+++ airfoil/adt_calc_kernel.cpp
@@ -11,13 +11,13 @@
                                                                                   
 void op_x86_adt_calc(                                                             
   int    blockIdx,                                                                
-  float *ind_arg0, int *ind_arg0_maps,                                            
+  double *ind_arg0, int *ind_arg0_maps,                                            
   short *arg0_maps,                                                               
   short *arg1_maps,                                                               
   short *arg2_maps,                                                               
   short *arg3_maps,                                                               
-  float *arg4,                                                                    
-  float *arg5,                                                                    
+  double *arg4,                                                                    
+  double *arg5,                                                                    
   int   *ind_arg_sizes,                                                           
   int   *ind_arg_offs,                                                            
   int    block_offset,                                                            
@@ -29,7 +29,7 @@ void op_x86_adt_calc(
                                                                                   
                                                                                   
   int   *ind_arg0_map, ind_arg0_size;                                  
-  float *ind_arg0_s;                                                   
+  double *ind_arg0_s;                                                   
   int    nelem, offset_b;                                              
                                                                                   
   char shared[64000];                                                  
@@ -49,7 +49,7 @@ void op_x86_adt_calc(
     // set shared memory pointers                                                 
                                                                                   
     int nbytes = 0;                                                               
-    ind_arg0_s = (float *) &shared[nbytes];                                       
+    ind_arg0_s = (double *) &shared[nbytes];                                       
   }                                                                               
                                                                                   
   __syncthreads(); // make sure all of above completed                            
@@ -131,13 +131,13 @@ void op_par_loop_adt_calc(char const *name, op_set set,
 #pragma omp parallel for                                                          
     for (int blockIdx=0; blockIdx<nblocks; blockIdx++)                            
      op_x86_adt_calc( blockIdx,                                                   
-       (float *)arg0.dat, (*Plan).ind_maps[0],                                    
+       (double *)arg0.dat, (*Plan).ind_maps[0],                                    
        (*Plan).maps[0],                                                           
        (*Plan).maps[1],                                                           
        (*Plan).maps[2],                                                           
        (*Plan).maps[3],                                                           
-       (float *)arg4.dat,                                                         
-       (float *)arg5.dat,                                                         
+       (double *)arg4.dat,                                                         
+       (double *)arg5.dat,                                                         
        (*Plan).ind_sizes,                                                         
        (*Plan).ind_offs,                                                          
        block_offset,
